+++ src/de/mhlz/MIPSInterpreter/Operations/IType/BEQOperation.java
@@ -7,7 +7,7 @@ import de.mhlz.MIPSInterpreter.Operations.IOperation;
  * Created by mischa on 11.02.14.
  */
 public class BEQOperation extends IOperation {
-	public BEQOperation(Interpreter i, int rs, int rt, short immediate) {
+	public BEQOperation(Interpreter i, int rs, int rt, int immediate) {
 		super(i, rs, rt, immediate);
 	}
